+++ iOS/BugsnagTests/BugsnagEventFromKSCrashReportTest.m
@@ -53,7 +53,7 @@
 
 - (void)testAddMetadataAddsNewTab {
     NSDictionary *metadata = @{@"color" : @"blue", @"beverage" : @"tea"};
-    [self.report addMetadata:metadata toTabWithName:@"user prefs"];
+    [self.report addMetadata:metadata toSectionNamed:@"user prefs"];
     NSDictionary *prefs = self.report.metadata[@"user prefs"];
     XCTAssertEqual(@"blue", prefs[@"color"]);
     XCTAssertEqual(@"tea", prefs[@"beverage"]);
@@ -62,9 +62,9 @@
 
 - (void)testAddMetadataMergesExistingTab {
     NSDictionary *oldMetadata = @{@"color" : @"red", @"food" : @"carrots"};
-    [self.report addMetadata:oldMetadata toTabWithName:@"user prefs"];
+    [self.report addMetadata:oldMetadata toSectionNamed:@"user prefs"];
     NSDictionary *metadata = @{@"color" : @"blue", @"beverage" : @"tea"};
-    [self.report addMetadata:metadata toTabWithName:@"user prefs"];
+    [self.report addMetadata:metadata toSectionNamed:@"user prefs"];
     NSDictionary *prefs = self.report.metadata[@"user prefs"];
     XCTAssertEqual(@"blue", prefs[@"color"]);
     XCTAssertEqual(@"tea", prefs[@"beverage"]);
@@ -73,25 +73,31 @@
 }
 
 - (void)testAddAttributeAddsNewTab {
-    [self.report addAttribute:@"color"
-                    withValue:@"blue"
-                toTabWithName:@"prefs"];
+    [self.report addMetadataToSectionNamed:@"prefs"
+                                       key:@"color"
+                                     value:@"blue"];
     NSDictionary *prefs = self.report.metadata[@"prefs"];
     XCTAssertEqual(@"blue", prefs[@"color"]);
 }
 
 - (void)testAddAttributeOverridesExistingValue {
-    [self.report addAttribute:@"color" withValue:@"red" toTabWithName:@"prefs"];
-    [self.report addAttribute:@"color"
-                    withValue:@"blue"
-                toTabWithName:@"prefs"];
+    [self.report addMetadataToSectionNamed:@"prefs"
+                                       key:@"color"
+                                     value:@"red"];
+    [self.report addMetadataToSectionNamed:@"prefs"
+                                       key:@"color"
+                                     value:@"blue"];
     NSDictionary *prefs = self.report.metadata[@"prefs"];
     XCTAssertEqual(@"blue", prefs[@"color"]);
 }
 
 - (void)testAddAttributeRemovesValue {
-    [self.report addAttribute:@"color" withValue:@"red" toTabWithName:@"prefs"];
-    [self.report addAttribute:@"color" withValue:nil toTabWithName:@"prefs"];
+    [self.report addMetadataToSectionNamed:@"prefs"
+                                       key:@"color"
+                                     value:@"red"];
+    [self.report addMetadataToSectionNamed:@"prefs"
+                                       key:@"color"
+                                     value:nil];
     NSDictionary *prefs = self.report.metadata[@"prefs"];
     XCTAssertNil(prefs[@"color"]);
 }